requirements.txt
gradio
tensorflow
tf_keras
Pillow
numpy